assignment 2
We define a Node class to represent each node in the linked list. Each node has a data attribute to store the node's data, and a next attribute to store a pointer to the next node in the list.

The getMiddle() function takes the head pointer of the linked list as input. We initialize two pointers, slow and fast, both pointing to the head of the list. The fast pointer moves twice as fast as the slow pointer. When fast reaches the end of the list, slow will be pointing to the middle node.

If the number of nodes is even, the slow pointer will be pointing to the second middle node. This is because we're moving the fast pointer twice as fast as the slow pointer, so when the fast pointer reaches the end of the list, the slow pointer will be pointing to the middle node if there are an odd number of nodes, and to the second middle node if there are an even number of nodes.

Finally, we return the data attribute of the middle node (or second middle node if the number of nodes is even).